Permission

+++ app/Http/Controllers/MahasiswaController.php
@@ -33,19 +33,33 @@ class MahasiswaController extends Controller
             return DataTables::of($mhs)
                 ->addIndexColumn()
                 ->addColumn('action', function ($row) {
+                    $user = auth()->user();
+
+                    $editBtn = '';
+                    $deleteBtn = '';
+                    $bayarBtn = '';
                     $showBtn = '<a href="' . route('mahasiswa.show', $row->id) . '" class="btn icon btn-sm btn-info" title="Detail"><i class="bi bi-eye"></i></a>';
-                    $editBtn = '<a href="' . route('mahasiswa.edit', $row->id) . '" class="btn icon btn-sm btn-warning" title="Edit"><i class="bi bi-pencil-square"></i></a>';
-                    $deleteBtn = '<button class="btn icon btn-sm btn-danger delete-btn" title="Delete" onclick="deleteMahasiswa(' . $row->id . ')">
+
+                    if ($user->can('update_mahasiswa')) {
+                        $editBtn = '<a href="' . route('mahasiswa.edit', $row->id) . '" class="btn icon btn-sm btn-warning" title="Edit"><i class="bi bi-pencil-square"></i></a>';
+                    }
+
+                    if ($user->can('delete_mahasiswa')) {
+                        $deleteBtn = '<button class="btn icon btn-sm btn-danger delete-btn" title="Delete" onclick="deleteMahasiswa(' . $row->id . ')">
                                     <i class="bi bi-trash"></i>
                                 </button>';
-
-                    // Only show the "Bayar" button if the student's status is 0 (nonaktif)
-                    $bayarBtn = '';
-                    if ($row->status == 0) {
-                        $bayarBtn = '<button class="btn icon btn-sm btn-success ms-1" title="Bayar" data-semester="' . $row->semester_berjalan . '" data-mahasiswa-id="' . $row->id . '" data-program-studi-id="' . $row->program_studi_id . '" data-nama="' . $row->nama . '" data-bs-toggle="modal" data-bs-target="#bayarModal" onclick="openBayarModal(' . $row->id . ')">
-                                    <i class="bi bi-cash"></i>
-                                </button>';
                     }
+
+                    if ($user->can('update_mahasiswa')) {
+                        // Only show the "Bayar" button if the student's status is 0 (nonaktif)
+                        $bayarBtn = '';
+                        if ($row->status == 0) {
+                            $bayarBtn = '<button class="btn icon btn-sm btn-success ms-1" title="Bayar" data-semester="' . $row->semester_berjalan . '" data-mahasiswa-id="' . $row->id . '" data-program-studi-id="' . $row->program_studi_id . '" data-nama="' . $row->nama . '" data-bs-toggle="modal" data-bs-target="#bayarModal" onclick="openBayarModal(' . $row->id . ')">
+                <i class="bi bi-cash"></i>
+            </button>';
+                        }
+                    }
+
                     return $showBtn . ' ' . $editBtn . ' ' . $deleteBtn . '' . $bayarBtn;
                 })
                 ->rawColumns(['action']) // Ensures that HTML code for the action buttons is rendered correctly
